usbwidget: Centralized Build Variables
Getting a bit tired of updating multiple files when build-variables
change. It now works as follows:

Bldvars.mif contains the various build-variables.
It is included by the Makefile and the Bldvars.h, Bldvars.rexx and the
module-definition (*.def) files are generated from it. Bldlevel.h
includes Bldvars.h and the *.lvl files construct the BLDLEVEL info that is
embedded in the Daemon and Widgets modules. Modifying Bldvars.mif will
cause both modules to be rebuilt with the updated values.