usbwidget: Added new 'Widget' class
We're working towards an Object Oriented anatomy of the UsbWidget.
By also encapsulating the XCenter Widget in a class we'll have more
flexibility. The goal is to allow for dynamic loading and unloading
a widget implementation so the WPS does not have to be restarted to
activate the new or changed functionality.